requirements.txt
requests
pysocks
cryptography